+++ Sources/DomainArchitectureMacros/Extensions/AttributeListSyntax+Extensions.swift
@@ -1,0 +1,32 @@
+import SwiftSyntax
+
+extension AttributeListSyntax.Element {
+    var availability: AttributeListSyntax.Element? {
+        switch self {
+        case .attribute(let attribute):
+            if let availability = attribute.availability {
+                return .attribute(availability)
+            }
+        case .ifConfigDecl(let ifConfig):
+            if let availability = ifConfig.availability {
+                return .ifConfigDecl(availability)
+            }
+        }
+        return nil
+    }
+}
+
+extension AttributeListSyntax {
+    var availability: AttributeListSyntax? {
+        var elements = [AttributeListSyntax.Element]()
+        for element in self {
+            if let availability = element.availability {
+                elements.append(availability)
+            }
+        }
+        if elements.isEmpty {
+            return nil
+        }
+        return AttributeListSyntax(elements)
+    }
+}
