+++ Sources/DomainArchitectureMacros/Extensions/AttributeSyntax+Extensions.swift
@@ -1,0 +1,11 @@
+import SwiftSyntax
+
+extension AttributeSyntax {
+    var availability: AttributeSyntax? {
+        if attributeName.identifier == "available" {
+            return self
+        } else {
+            return nil
+        }
+    }
+}
