+++ Sources/DomainArchitectureMacros/Extensions/String+Extensions.swift
@@ -1,0 +1,5 @@
+extension [String] {
+    var moduleQualified: Self {
+        self.flatMap { [$0, "DomainArchitecture.\($0)"] }
+    }
+}
