+++ Sources/DomainArchitectureMacros/Extensions/TypeSyntax+Extensions.swift
@@ -1,0 +1,15 @@
+import SwiftSyntax
+
+extension TypeSyntax {
+    var identifier: String? {
+        for token in tokens(viewMode: .all) {
+            switch token.tokenKind {
+            case .identifier(let identifier):
+                return identifier
+            default:
+                break
+            }
+        }
+        return nil
+    }
+}
