+++ Sources/DomainArchitectureMacros/Plugins/InteractorMacro.swift
@@ -44,7 +44,7 @@ extension InteractorMacro: MemberAttributeMacro {
     ) throws -> [AttributeSyntax] {
         let macroGenerics = node.attributeName.as(IdentifierTypeSyntax.self)?.genericArgumentClause?.arguments
         if let macroGenerics {
-            guard macroGenerics.count == 2 else {
+            if macroGenerics.count > 2 {
                 context
                     .diagnose(
                         Diagnostic(
@@ -126,11 +126,20 @@ extension InteractorMacro: MemberAttributeMacro {
                     return []
                 }
             }
+
+            let builderArguments: TokenSyntax =
+                if let macroGenerics {
+                    .identifier("DomainArchitecture.InteractorBuilder<\(macroGenerics)>")
+                } else if genericArgs.count == 1 {
+                    .identifier(
+                        "DomainArchitecture.InteractorBuilder<\(genericArgs.description).State, \(genericArgs.description).Action>"
+                    )
+                } else {
+                    .identifier("DomainArchitecture.InteractorBuilder<\(genericArgs)>")
+                }
             return [
                 AttributeSyntax(
-                    attributeName: IdentifierTypeSyntax(
-                        name: .identifier("DomainArchitecture.InteractorBuilder<\(macroGenerics ?? genericArgs)>")
-                    )
+                    attributeName: IdentifierTypeSyntax(name: builderArguments)
                 )
             ]
         }
@@ -163,60 +172,50 @@ extension InteractorMacro: MemberMacro {
             // TODO: - Diagnostic?
             return []
         }
-        guard let generics = attrName.genericArgumentClause,
+
+        if let generics = attrName.genericArgumentClause,
             generics.arguments.count == 2
-        else {
-            context
-                .diagnose(
-                    Diagnostic(
-                        node: node.attributeName,
-                        message: MacroExpansionErrorMessage(
-                            """
-                            Only 2 generic arguments should be applied the @Interactor macro. \
-                            One for the Interactor's state type and one for its action type. 
-                            """
-                        )
-                    )
+        {
+            let argumentsArray = generics
+                .arguments
+                .compactMap { $0.argument.as(IdentifierTypeSyntax.self) }
+            let domainStateType = argumentsArray[0].name.text
+            let eventType = argumentsArray[1].name.text
+            var decls: [DeclSyntax] = []
+
+            handleTypeAlias(
+                existingTypeAliases,
+                context: context,
+                aliasType: .init(
+                    rawValue: "DomainState",
+                    typeName: domainStateType
                 )
+            ) {
+                decls.append(
+                    """
+                    typealias DomainState = \(raw: domainStateType)
+                    """
+                )
+            }
+
+            handleTypeAlias(
+                existingTypeAliases,
+                context: context,
+                aliasType: .init(
+                    rawValue: "Action",
+                    typeName: eventType
+                )
+            ) {
+                decls.append(
+                    """
+                    typealias Action = \(raw: eventType)
+                    """
+                )
+            }
+            return decls
+        } else {
             return []
         }
-        let argumentsArray = generics
-            .arguments
-            .compactMap { $0.argument.as(IdentifierTypeSyntax.self) }
-        let domainStateType = argumentsArray[0].name.text
-        let eventType = argumentsArray[1].name.text
-        var decls: [DeclSyntax] = []
-
-        handleTypeAlias(
-            existingTypeAliases,
-            context: context,
-            aliasType: .init(
-                rawValue: "DomainState",
-                typeName: domainStateType
-            )
-        ) {
-            decls.append(
-                """
-                typealias DomainState = \(raw: domainStateType)
-                """
-            )
-        }
-
-        handleTypeAlias(
-            existingTypeAliases,
-            context: context,
-            aliasType: .init(
-                rawValue: "Action",
-                typeName: eventType
-            )
-        ) {
-            decls.append(
-                """
-                typealias Action = \(raw: eventType)
-                """
-            )
-        }
-        return decls
     }
 
     private static func handleTypeAlias<C: MacroExpansionContext>(
@@ -245,115 +244,5 @@ extension InteractorMacro: MemberMacro {
     private struct TypeAliasType {
         let rawValue: String
         let typeName: String
-    }
-}
-
-extension [String] {
-    var moduleQualified: Self {
-        self.flatMap { [$0, "DomainArchitecture.\($0)"] }
-    }
-}
-
-extension AttributeSyntax {
-    var availability: AttributeSyntax? {
-        if attributeName.identifier == "available" {
-            return self
-        } else {
-            return nil
-        }
-    }
-}
-
-extension TypeSyntax {
-    var identifier: String? {
-        for token in tokens(viewMode: .all) {
-            switch token.tokenKind {
-            case .identifier(let identifier):
-                return identifier
-            default:
-                break
-            }
-        }
-        return nil
-    }
-}
-
-extension AttributeListSyntax.Element {
-    var availability: AttributeListSyntax.Element? {
-        switch self {
-        case .attribute(let attribute):
-            if let availability = attribute.availability {
-                return .attribute(availability)
-            }
-        case .ifConfigDecl(let ifConfig):
-            if let availability = ifConfig.availability {
-                return .ifConfigDecl(availability)
-            }
-        }
-        return nil
-    }
-}
-
-extension AttributeListSyntax {
-    var availability: AttributeListSyntax? {
-        var elements = [AttributeListSyntax.Element]()
-        for element in self {
-            if let availability = element.availability {
-                elements.append(availability)
-            }
-        }
-        if elements.isEmpty {
-            return nil
-        }
-        return AttributeListSyntax(elements)
-    }
-}
-
-extension IfConfigDeclSyntax {
-    var availability: IfConfigDeclSyntax? {
-        var elements = [IfConfigClauseListSyntax.Element]()
-        for clause in clauses {
-            if let availability = clause.availability {
-                if elements.isEmpty {
-                    elements.append(availability.clonedAsIf)
-                } else {
-                    elements.append(availability)
-                }
-            }
-        }
-        if elements.isEmpty {
-            return nil
-        } else {
-            return with(\.clauses, IfConfigClauseListSyntax(elements))
-        }
-    }
-}
-
-extension IfConfigClauseSyntax {
-    var availability: IfConfigClauseSyntax? {
-        if let availability = elements?.availability {
-            return with(\.elements, availability)
-        } else {
-            return nil
-        }
-    }
-
-    var clonedAsIf: IfConfigClauseSyntax {
-        detached.with(\.poundKeyword, .poundIfToken())
-    }
-}
-
-extension IfConfigClauseSyntax.Elements {
-    var availability: IfConfigClauseSyntax.Elements? {
-        switch self {
-        case .attributes(let attributes):
-            if let availability = attributes.availability {
-                return .attributes(availability)
-            } else {
-                return nil
-            }
-        default:
-            return nil
-        }
     }
 }
